ai-service/requirements.txt
flask
requests
python-dotenv
groq
flask-limiter